requirements.txt
certifi==2021.5.30
click==8.0.4
dataclasses==0.8
dateutils==0.6.12
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
